requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgi-correlation-id==4.3.4
attrs==25.4.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.2
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
frozenlist==1.8.0
fsspec==2025.10.0
gcsfs==2025.10.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.3
oauthlib==3.3.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
alembic==1.14.0
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xlrd==2.0.2
yarl==1.22.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
alembic==1.14.0
